extract kubeconfig merge to a separate script to make it optional

+++ scripts/deploy-bootstrap-cluster.sh
@@ -1,10 +1,22 @@
 #!/bin/bash
-
-set -eoux pipefail
-
-workdir=$(pwd)
+set -eou pipefail
 
 # For more details please check docs/bootstrap-and-pivot.md doc in this repo
+
+
+# Setup config environment variables, and AWS_B64ENCODED_CREDENTIALS
+# run `clusterctl generate cluster --list-variables aws` to get the list of variables
+if [ -z "$AWS_CONTROL_PLANE_MACHINE_TYPE" ] || \
+   [ -z "$AWS_NODE_MACHINE_TYPE" ] || \
+   [ -z "$AWS_SSH_KEY_NAME" ] || \
+   [ -z "$KUBERNETES_VERSION" ] || \
+   [ -z "$AWS_B64ENCODED_CREDENTIALS" ]; then
+  exit 1
+fi
+
+set -x
+
+workdir=$(pwd)
 
 trap "rm -f bootstrap.yaml" EXIT
 cat > bootstrap.yaml << EOF
@@ -18,17 +30,6 @@ EOF
 kind create cluster --config bootstrap.yaml
 
 clusterctl init --infrastructure aws
-
-# Setup config environment variables, and AWS_B64ENCODED_CREDENTIALS
-# run `clusterctl generate cluster --list-variables aws` to get the list of variables
-if [ -z "$AWS_CONTROL_PLANE_MACHINE_TYPE" ] || \
-   [ -z "$AWS_NODE_MACHINE_TYPE" ] || \
-   [ -z "$AWS_SSH_KEY_NAME" ] || \
-   [ -z "$KUBERNETES_VERSION" ] || \
-   [ -z "$AWS_B64ENCODED_CREDENTIALS" ]; then
-  # unreachable code due to set -u. needs better 'trapping' to provide an error message
-  exit 1
-fi
 
 set +e
 while ! kubectl get clusters; do
@@ -45,13 +46,14 @@ echo
 
 # applying cluster manifests immediatelly will fail because components webhooks are not yet ready to serve traffic
 # it is easier to retry applying rather than checking on each component individually
-retries=5
+retries=8
 set +e
 kubectl apply -f $workdir/mgmt.yaml 2>/dev/null
 while [ $? -ne 0 ]; do
   echo Failed to apply cluster config, re-trying
-  sleep 10
-  [[ $retries -eq 0 ]] && echo "Failed to apply clsuter config, aborting." && exit 1
+  sleep 15
+  # if retries are exhausted, there might be a genuine error, run the command one more time without swallowing the  error
+  [[ $retries -eq 0 ]] && kubectl apply -f $workdir/mgmt.yaml && echo "Failed to apply cluster config, aborting." && exit 1
   ((retries--))
   kubectl apply -f $workdir/mgmt.yaml 2>/dev/null
 done
@@ -67,32 +69,24 @@ done
 sleep 15 # wait for `k get secret mgmt-kubeconfig`
 clusterctl get kubeconfig mgmt > $workdir/target-mgmt.kubeconfig
 
-# backup previous kubeconfig, this is also needed for merge: can't read and redirect to the same place in one command
-temp_kubeconfig=$HOME/.kube/config-$(date +%F_%H_%M_%S)
-cp $HOME/.kube/config $temp_kubeconfig
-KUBECONFIG=$workdir/target-mgmt.kubeconfig:$temp_kubeconfig kubectl config view --raw=true --merge=true > $HOME/.kube/config
+# check out ./scripts/merge-kubeconfig.sh to merge all kubeconfigs into default ~/.kube/config file (preserving already existing configs)
 
-##############
+
 ############## ------ on AWS mgmt cluster ------
-##############
 
-kubectl config use-context mgmt-admin@mgmt
+sleep 45  # something is still not ready, wait
+kubectl --kubeconfig $workdir/target-mgmt.kubeconfig apply -f https://docs.projectcalico.org/v3.21/manifests/calico.yaml
 
-sleep 30  # something is still not ready, wait
-kubectl apply -f https://docs.projectcalico.org/v3.21/manifests/calico.yaml
-
-clusterctl init --infrastructure aws
+clusterctl init --infrastructure aws --kubeconfig $workdir/target-mgmt.kubeconfig --kubeconfig-context mgmt-admin@mgmt
 
 set +e
-set -x
-while ! kubectl get clusters; do
+while ! kubectl --kubeconfig=$workdir/target-mgmt.kubeconfig get clusters; do
   sleep 15
 done
 set -e
-set +x
 echo \"No resources found in default namespace\" expected
 
-clusterctl move --to-kubeconfig=./target-mgmt.kubeconfig
+clusterctl move --kubeconfig=$workdir/target-mgmt.kubeconfig --kubeconfig-context mgmt-admin@mgmt --to-kubeconfig=./target-mgmt.kubeconfig
 
 # kind delete cluster
 # and now what? how do you manage the permanent management cluster? keep it now for simplicity
